Fix Vptr.disp size to fit MemoryRef.disp
There is an implicit ctor from one to another:
  /* implicit */ Vptr(MemoryRef m, Segment s = DS)

However Vptr.disp doesn't fit MemoryRef.disp:
MemoryRef:  intptr_t
Vptr:       int32_t

Therefore the following was failing:
Vptr p = baseless(inst.s.r.disp);

inst:
$51 = (HPHP::jit::loadqp &) @0x3fff7ab20450: {
  s = {
    r = {
      disp = 0x10039a00000
    }
  },
  d = {
    rn = 0x41
  }
}

Resulting p:
$54 = {
  base = {
    rn = 0xffffffff
  },
  index = {
    rn = 0xffffffff
  },
  scale = 0x1,
  seg = 0x0,
  disp = 0x39a00000
}

Disp attribute lost precision and there were no warnings when building it.

In PPC64 the promotions from Immed to Immed64 needed to be done as all
'disp' attributes from Vptr doesn't fit in Immed anymore.

+++ hphp/runtime/vm/jit/vasm-reg.h
@@ -265,7 +265,7 @@ struct Vptr {
   Vreg64 index;     // optional
   uint8_t scale;    // 1,2,4,8
   Segment seg{DS};  // DS, FS or GS
-  int32_t disp;
+  intptr_t disp;    // big enough to fit a MemoryRef.disp type
 };
 
 Vptr operator+(Vptr lhs, int32_t d);
